- Syncing with release - Added an export function for CSV (Exporting is unfinished) - Altering calculations slightly based on simulations (more to come)

+++ start_ui.py
@@ -6,7 +6,6 @@
 import sys, os, time
 import common
 from PyQt5.QtWidgets import QApplication, QStyleFactory
-from PyQt5 import QtGui, QtCore
 from FrmMain import Frm_Main
 
 
@@ -16,17 +15,21 @@ relative_path_covnert = common.relative_path_covnert
 
 get_dark_palette = Qt_common.get_dark_palette
 setIcon = Qt_common.setIcon
-MAXIMUM_LOGFILE_SIZE = 500 *1024
+MAXIMUM_LOGFILE_SIZE = 500 * 1024
 
 if __name__ == "__main__":
     log_path = relative_path_covnert('LOG.log')
-    file_size = os.stat(log_path).st_size
-    if file_size > MAXIMUM_LOGFILE_SIZE:
-        with open(log_path, 'rb') as f:
-            file_contents = f.read()
-        file_contents = file_contents[file_size-MAXIMUM_LOGFILE_SIZE:]
-        with open(log_path, 'wb') as f:
-            f.write(file_contents)
+    if os.path.isfile(log_path):
+        file_size = os.stat(log_path).st_size
+        if file_size > MAXIMUM_LOGFILE_SIZE:
+            with open(log_path, 'rb') as f:
+                file_contents = f.read()
+            file_contents = file_contents[file_size-MAXIMUM_LOGFILE_SIZE:]
+            with open(log_path, 'wb') as f:
+                f.write(file_contents)
+    elif os.path.isdir(log_path):
+        print 'Log file cannot be a directory.'
+        sys.exit(1)
     tee = utils.Tee(log_path, 'a')
     print 'Starting: ' + str(time.time())
     try:
